+++ psutil/_pslinux.py
@@ -1333,15 +1333,14 @@ class Process(object):
 
     @wrap_exceptions
     def status(self):
-        with open_binary("%s/%s/status" % (self._procfs_path, self.pid)) as f:
-            for line in f:
-                if line.startswith(b"State:"):
-                    letter = line.split()[1]
-                    if PY3:
-                        letter = letter.decode()
-                    # XXX is '?' legit? (we're not supposed to return
-                    # it anyway)
-                    return PROC_STATUSES.get(letter, '?')
+        with open_binary("%s/%s/stat" % (self._procfs_path, self.pid)) as f:
+            data = f.read()
+        letter = data[data.rfind(b')') + 2:][0:1]
+        if PY3:
+            letter = letter.decode()
+        # XXX is '?' legit? (we're not supposed to return
+        # it anyway)
+        return PROC_STATUSES.get(letter, '?')
 
     @wrap_exceptions
     def open_files(self):